Action cleanup
Moved Vector3Compare to Logic category (as is defined in the script)

Renamed GetMousePosition (GUI category) to GetMousePositionAdvanced to
avoid name dupe with older action GetMousePosition (Input category)

Moved GameObjectIsActiveInHierarchy to U4.x GitRep due to exception.
Action named IsActive works for 3.x and 4.x and is similar, which
remains in the 3.x rep.